my commits

+++ 0x00-hello_world/6-size.c
@@ -1,14 +1,19 @@
 #include <stdio.h>
+
 /**
  * main -entry point
+ * description: displays size of various data types of the machine it is compiled on
  * Return: always 0(success)
  */
+
 int main(void)
 {
-printf("Size of a char: %i byte(s)\n", sizeof(char));
-printf("Size of an int: %i byte(s)\n", sizeof(int));
-printf("Size of a long int: %i bytes(s)\n", sizeof(long int));
-printf("Size of a long long int: %i byte(s)\n", sizeof(long long int));
-printf("Size of a float: %i bytes(s)", sizeof(float));
+
+printf("Size of a char: %i byte(s)\n",(int) sizeof(char));
+printf("Size of an int: %i byte(s)\n",(int) sizeof(int));
+printf("Size of a long int: %i bytes(s)\n",(int) sizeof(long int));
+printf("Size of a long long int: %i byte(s)\n",(int) sizeof(long long int));
+printf("Size of a float: %i bytes(s)", (int)sizeof(float));
+
 return (0);
 }
